Newsletter-main/requirements.txt
fastapi
uvicorn
python-dotenv
requests
arxiv
langchain
langchain-community
langchain-groq
langchain-tavily
-e .
